% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolescModel.R
\name{adolescModel}
\alias{adolescModel}
\title{Creates the adolescence model from individual cross-secional data based current economic status and educational level.}
\usage{
adolescModel(cs.data.all)
}
\arguments{
\item{cs.data.all}{Cross-sectional individual data.}
}
\description{
Creates the adolescence model from individual cross-secional data based current economic status and educational level.
}
