% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolescModel.R
\name{adolescSeq}
\alias{adolescSeq}
\title{Predicts school type.}
\usage{
adolescSeq(schoolTypeAtAge16, n)
}
\arguments{
\item{schoolTypeAtAge16}{School type probabilities at age 16}

\item{n}{Sample zize.}
}
\description{
Predicts school type.
}
\seealso{
\code{\link{adolescModel}} \code{\link{predictSchoolTypeProb}}
}
