% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovChain.R
\name{asMarkovChain}
\alias{asMarkovChain}
\title{Determines the matrix of transition rates from the sequence data and creates a Markov chain.}
\usage{
asMarkovChain(data.seq, data.scodes, brewer.pal.name = NULL)
}
\arguments{
\item{data.seq}{data.seq}

\item{data.scodes}{data.scodes}

\item{brewer.pal.name}{Name of the RColorBrewer palette to use. TraMineR uses the following: If number of states <= 8 then 'Accent'; if between 8 and 12, then 'Set3'; if greater than 12 you have to specify your own palette.}
}
\description{
Determines the matrix of transition rates from the sequence data and creates a Markov chain.
}
