% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parLifeTraj.R
\name{determineFullTraj}
\alias{determineFullTraj}
\title{Runs the trajectory simulation in parallel.}
\usage{
determineFullTraj(INDIV_SUBJID, INDIV_AGE, INDIV_SEX, INDIV_EDU, INDIV_ACT)
}
\description{
Runs the trajectory simulation in parallel.
}
