% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineSeqTree.R
\name{determineSeqTree}
\alias{determineSeqTree}
\title{Determines the sequence tree.}
\usage{
determineSeqTree(data.rshp.subset, dist.method = "OM", dist.indel = 1)
}
\arguments{
\item{data.rshp.subset}{}

\item{dist.method}{See \link[TraMineR]{seqdist} for details.}

\item{dist.indel}{See \link[TraMineR]{seqdist} for details.}
}
\description{
Determines the sequence tree.
}
