% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTree.R
\name{findNodeById}
\alias{findNodeById}
\title{Finds a specific node in the sequence dissimilarity tree.}
\usage{
findNodeById(node, node.id)
}
\arguments{
\item{node}{Node in sequence dissimilarity tree}

\item{node.id}{node.id}
}
\description{
Finds a specific node in the sequence dissimilarity tree.
}
