% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrossSectEdu.R
\name{getCrossSectEdu}
\alias{getCrossSectEdu}
\title{Determine cross-sectional education data in terms of highest attained educational level and 
current education type.}
\usage{
getCrossSectEdu(cs.data.all, country.filter = NULL)
}
\description{
Determine cross-sectional education data in terms of highest attained educational level and 
current education type.
}
