% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExposureData.R
\name{getExposureData}
\alias{getExposureData}
\title{Reads individual-specific daily exposure data from a given path.}
\usage{
getExposureData(INDIV_SUBJID, stressors, PATH)
}
\description{
It is assumed that there is a subfolder for each individual stressor.
Furthermore, it is assumed that one separate file per individual is given.
}
