% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTree.R
\name{getNodeId}
\alias{getNodeId}
\title{Determines corresponding leaf node in the sequence dissimilarity tree for a
given set of characteristics (age, sex, etc.).}
\usage{
getNodeId(df, age, sex)
}
\arguments{
\item{age}{age}

\item{sex}{sex}
}
\description{
Determines corresponding leaf node in the sequence dissimilarity tree for a
given set of characteristics (age, sex, etc.).
}
