% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTree.R
\name{isLeafNode}
\alias{isLeafNode}
\title{Determines whether a node in the sequence dissimilarity tree is a leaf node.
Only sequences within leaf nodes will be considered as they are not split further.}
\usage{
isLeafNode(node)
}
\arguments{
\item{node}{Node in sequence dissimilarity tree}
}
\description{
Determines whether a node in the sequence dissimilarity tree is a leaf node.
Only sequences within leaf nodes will be considered as they are not split further.
}
