% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parLifecourseExposure.R
\name{par_lifeCourseExposure}
\alias{par_lifeCourseExposure}
\title{Prallel estimation of life-long exposure of the given individual.}
\usage{
par_lifeCourseExposure(i, num.cores = NUM_CORES, cluster.type = "PSOCK")
}
\arguments{
\item{i}{Index of individual.}

\item{num.cores}{Number of cores used.}

\item{cluster.type}{Type of cluster, e.g. MPI. Defaults to PSOCK.}
}
\description{
Prallel estimation of life-long exposure of the given individual.
}
