% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parLifeTraj.R
\name{par_lifeTrajSim}
\alias{par_lifeTrajSim}
\title{Simulation of the lifecourse trajectory of the given individual.}
\usage{
par_lifeTrajSim(sim.id, INDIV_AGE, INDIV_SEX, INDIV_ACT, INDIV_EDU)
}
\description{
Simulation of the lifecourse trajectory of the given individual.
}
