% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovChain.R
\name{plotMarkovChain}
\alias{plotMarkovChain}
\title{Plots the Markov chain determined from the transition matrix.}
\usage{
plotMarkovChain(net, edge.weight.mult = 5, arrow.size = 0.3)
}
\arguments{
\item{net}{Markov chain derived from calling asMarkovChain function.}

\item{edge.weight.mult}{Multiply the original edge weight by this factor.}

\item{arrow.size}{Size of the arrow tips.}
}
\description{
Plots the Markov chain determined from the transition matrix.
}
\seealso{
\code{\link{asMarkovChain}}
}
