% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolescModel.R
\name{predictSchoolTypeProb}
\alias{predictSchoolTypeProb}
\title{Predicts school type.}
\usage{
predictSchoolTypeProb(sex, activity, income, model)
}
\arguments{
\item{sex}{sex}

\item{activity}{activity}

\item{income}{income}

\item{model}{Adolescence model}
}
\description{
Predicts school type.
}
\seealso{
\code{\link{adolescModel}}
}
