% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHouseholdData.R
\name{readHouseholdData}
\alias{readHouseholdData}
\title{Reads household data from cross-sectional EU-SILC data}
\usage{
readHouseholdData(H.files.cs)
}
\arguments{
\item{H.files.cs}{List of H-files.}
}
\description{
Reads household data from cross-sectional EU-SILC data
}
