% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndividualData.R
\name{readIndividualData}
\alias{readIndividualData}
\title{Reads individual data from cross-sectional EU-SILC data (only age 16 and 17)}
\usage{
readIndividualData(P.files.cs, R.files.cs, cs.hh.data.all = NULL)
}
\arguments{
\item{P.files.cs}{List of P-files (should match R-files list)}

\item{cs.hh.data.all}{Optional household data that will be linked to the individual (household income etc.)}

\item{P.files.cs}{List of R-files (should match P-files list)}
}
\description{
Reads individual data from cross-sectional EU-SILC data (only age 16 and 17)
}
\seealso{
\code{\link{readHouseholdData}}
}
