% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndividualData.R
\name{readIndividualData2}
\alias{readIndividualData2}
\title{Reads individual data from cross-sectional EU-SILC data (only age 16 and 17)}
\usage{
readIndividualData2(P.files.cs, cs.hh.data.all = NULL)
}
\arguments{
\item{P.files.cs}{List of P-files}

\item{cs.hh.data.all}{Optional household data that will be linked to the individual (household income etc.)}
}
\description{
Reads individual data from cross-sectional EU-SILC data (only age 16 and 17)
}
\seealso{
\code{\link{readHouseholdData}}
}
