% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndividualSeq.R
\name{readIndividualSeq}
\alias{readIndividualSeq}
\title{Reads individual 4-year sequences from longitudinal EU-SILC data}
\usage{
readIndividualSeq(P.files.lon, lon.hh.data.all)
}
\arguments{
\item{P.files.lon}{List of P-files.}

\item{lon.hh.data.all}{Longitudinal household data.}
}
\description{
Reads individual 4-year sequences from longitudinal EU-SILC data
}
