% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulates life trajectories}
\usage{
simulate(df, st, indiv.age, indiv.sex, indiv.activity, AGE.STEP = 3)
}
\arguments{
\item{df}{df}

\item{st}{st}

\item{indiv.age}{indiv.age}

\item{indiv.sex}{indiv.sex}

\item{indiv.activity}{indiv.activity}

\item{AGE.STEP}{AGE.STEP}
}
\description{
Simulates life trajectories
}
