% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedSeqSample.R
\name{stratifiedSeqSample}
\alias{stratifiedSeqSample}
\title{Creates a stratified sample of sequences.}
\usage{
stratifiedSeqSample(data.rshp, factor = 0.1)
}
\arguments{
\item{data.rshp}{}

\item{factor}{The relative sample size (defaults to 0.1, i.e. 10%).}
}
\description{
Whenever there are too many sequences, difficulties may arise when the distance matrix is being calculated.
The problem is described in more detail:
http://stackoverflow.com/questions/15929936/problem-with-big-data-during-computation-of-sequence-distances-using-tramine
}
\details{
Here, a stratified sub-sampling approach is follows. This ensures that sequence distribution is maintained.
}
