% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTree.R
\name{traversePreOrder}
\alias{traversePreOrder}
\title{Traverses the sequence dissimilarity tree.}
\usage{
traversePreOrder(node, df = NULL, min.age = 16, max.age = 100,
  min.sex = 1, max.sex = 2)
}
\arguments{
\item{node}{Node in sequence dissimilarity tree}

\item{min.age}{min.age}

\item{max.age}{max.age}

\item{min.sex}{min.sex}

\item{max.sex}{max.sex}
}
\description{
Traverses the sequence dissimilarity tree.
}
