% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTree.R
\name{visitTreeNode}
\alias{visitTreeNode}
\title{Visits a node in the sequence dissimilarity tree}
\usage{
visitTreeNode(node, min.age, max.age, min.sex, max.sex)
}
\arguments{
\item{node}{Node in sequence dissimilarity tree}

\item{min.age}{min.age}

\item{max.age}{max.age}

\item{min.sex}{min.sex}

\item{max.sex}{max.sex}
}
\description{
Visits a node in the sequence dissimilarity tree
}
