% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parLifecourseExposure.R
\name{par_lifeCourseExposure}
\alias{par_lifeCourseExposure}
\title{Prallel estimation of life-long exposure of the given individual.}
\usage{
par_lifeCourseExposure(i)
}
\arguments{
\item{i}{Index of individual.}
}
\description{
Prallel estimation of life-long exposure of the given individual.
}
