% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulates life trajectories}
\usage{
simulate(st, indiv.age, indiv.sex, indiv.activity, age.step = 3)
}
\arguments{
\item{st}{Sequence tree.}

\item{indiv.age}{Age of the individual.}

\item{indiv.sex}{Sex of the individual.}

\item{indiv.activity}{Activity of the individual at the given age.}

\item{age.step}{Number of additional age steps per given sequence (defaults to 3 due to sequence length of 4)}
}
\description{
Simulates life trajectories
}
